global_chem_extensions/requirements.txt
numpy>=1.20
bokeh
plotly
global-chem
kaleido
rdkit-pypi
flask
pandas
sklearn
dimorphite_dl
partialsmiles
pysmiles
molvs
molpdf
deepsmiles
requests
selfies